refactored class naming
