+++ common/core/src/test/java/zingg/common/core/util/IDataReader.java
@@ -1,4 +1,4 @@
-package zingg.common.core.block.data;
+package zingg.common.core.util;
 
 import com.opencsv.exceptions.CsvException;
 
@@ -6,6 +6,6 @@ import java.io.IOException;
 import java.net.URISyntaxException;
 import java.util.List;
 
-public interface DataReader {
+public interface IDataReader {
     List<String[]> readDataFromSource(String source) throws IOException, CsvException, URISyntaxException;
 }
